Custom send implementation

+++ ShortDev.Microsoft.ConnectedDevices/Messages/Connection/ConnectionResponse.cs
@@ -46,7 +46,7 @@ public sealed class ConnectionResponse : ICdpPayload<ConnectionResponse>
         {
             Result = result,
             HmacSize = reader.ReadUInt16(),
-            Nonce = new(reader.ReadUInt16()),
+            Nonce = new(reader.ReadUInt64()),
             MessageFragmentSize = reader.ReadUInt32(),
             PublicKeyX = reader.ReadBytesWithLength().ToArray(),
             PublicKeyY = reader.ReadBytesWithLength().ToArray()
